ref(onboarding): inline SCM replay hook into OSS
Self-hosted Sentry doesn't register the Replay integration, so
Sentry.getReplay() returns undefined there and the existing early-return
already makes the hook a no-op. The HookStore indirection through gsApp
isn't needed.

Move the implementation back into static/app, drop the gsApp hook and
its registration, and remove the Hooks type entry.
