+++ static/app/utils/replays/useReplayForCriticalFlow.tsx
@@ -1,6 +1,7 @@
-import {HookStore} from 'sentry/stores/hookStore';
+import {useEffect, useState} from 'react';
+import * as Sentry from '@sentry/react';
 
-export interface UseReplayForCriticalFlowOptions {
+interface UseReplayForCriticalFlowOptions {
   /**
    * Tag value applied to the replay (and other events) for this flow,
    * so it can be filtered later in the Replays product.
@@ -13,25 +14,47 @@ export interface UseReplayForCriticalFlowOptions {
   enabled?: boolean;
   /**
    * Probability [0, 1] that a given mount forces a full session replay.
-   * Independent of getsentry's global session sample rate. Defaults to 1.0
+   * Independent of the global session sample rate. Defaults to 1.0
    * (force every user in the flow).
    */
   sampleRate?: number;
 }
 
-function useNoopReplayForCriticalFlow(_options: UseReplayForCriticalFlowOptions) {}
-
 /**
  * Force a session replay for a critical flow regardless of the global
  * sample rate, and tag it for later discovery.
  *
- * The implementation lives in getsentry (since OSS does not register the
- * Replay integration). When the gsApp hook is not registered, this is a
- * no-op, so self-hosted callers pay nothing.
+ * No-op when the Replay integration isn't registered (e.g. self-hosted,
+ * dev, acceptance tests, or before lazy init has finished).
  */
-export function useReplayForCriticalFlow(options: UseReplayForCriticalFlowOptions) {
-  const hook =
-    HookStore.get('react-hook:use-replay-for-critical-flow')[0] ??
-    useNoopReplayForCriticalFlow;
-  hook(options);
+export function useReplayForCriticalFlow({
+  flowName,
+  enabled = true,
+  sampleRate = 1,
+}: UseReplayForCriticalFlowOptions) {
+  const [shouldForce] = useState(() => Math.random() < sampleRate);
+
+  useEffect(() => {
+    if (!enabled || !shouldForce) {
+      return undefined;
+    }
+
+    const replay = Sentry.getReplay();
+    if (!replay) {
+      return undefined;
+    }
+
+    Sentry.setTag('critical_flow', flowName);
+
+    // flush() handles all three states per the SDK docstring:
+    //   - not yet recording -> starts a session replay
+    //   - buffer mode -> upgrades buffer to session
+    //   - session mode -> queued no-op
+    replay.flush();
+
+    return () => {
+      replay.flush();
+      Sentry.setTag('critical_flow', undefined);
+    };
+  }, [enabled, shouldForce, flowName]);
 }
